# Plugins/BTCPayServer.Plugins.SatoshiTickets/Resources/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SatoshiTickets Greenfield API",
    "description": "REST API for the SatoshiTickets BTCPay Server plugin. Provides endpoints for managing events, ticket types, tickets, orders, and check-ins.\n\nAll endpoints are scoped to a specific store and require a Greenfield API key with appropriate permissions.\n\n**Authentication:** Pass your API key in the `Authorization` header:\n```\nAuthorization: token YOUR_API_KEY\n```\n\nCreate an API key under **Account > Manage Account > API Keys** in BTCPay Server.",
    "version": "1.2.2",
    "contact": {
      "name": "SatoshiTickets",
      "url": "https://github.com/TChukwuleta/BTCPayServerPlugins"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1/stores/{storeId}/satoshi-tickets",
      "description": "Current BTCPay Server instance",
      "variables": {
        "storeId": {
          "default": "your-store-id",
          "description": "The store ID"
        }
      }
    }
  ],
...
  "paths": {
...
    "/events/{eventId}/ticket-types": {
      "get": {
        "tags": ["Ticket Types"],
        "summary": "List ticket types",
        "description": "Returns all ticket types for an event.",
        "operationId": "getTicketTypes",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sort field.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["Name", "Price"],
              "default": "Name"
            }
          },
          {
            "name": "sortDir",
            "in": "query",
            "description": "Sort direction.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of ticket types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketTypeData"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      },
      "post": {
        "tags": ["Ticket Types"],
        "summary": "Create ticket type",
        "description": "Creates a new ticket type (pricing tier) for an event. The first ticket type created automatically becomes the default.",
        "operationId": "createTicketType",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketTypeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket type created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketTypeData"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      }
    },
    "/events/{eventId}/ticket-types/{ticketTypeId}": {
      "get": {
        "tags": ["Ticket Types"],
        "summary": "Get ticket type",
        "description": "Returns a single ticket type.",
        "operationId": "getTicketType",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          },
          {
            "$ref": "#/components/parameters/ticketTypeId"
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket type details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketTypeData"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      },
      "put": {
        "tags": ["Ticket Types"],
        "summary": "Update ticket type",
        "description": "Updates an existing ticket type. If `isDefault` is set to `false` and no other type is default, this one remains default.",
        "operationId": "updateTicketType",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          },
          {
            "$ref": "#/components/parameters/ticketTypeId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketTypeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ticket type updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketTypeData"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      }
    },
    "/ticket-types/{ticketTypeId}": {
      "delete": {
        "tags": ["Ticket Types"],
        "summary": "Delete ticket type",
        "description": "Deletes a ticket type. If the deleted type was the default, another type is automatically assigned as default.",
        "operationId": "deleteTicketType",
        "parameters": [
          {
            "$ref": "#/components/parameters/ticketTypeId"
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket type deleted"
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      }
    },
    "/ticket-types/{ticketTypeId}/toggle": {
      "put": {
        "tags": ["Ticket Types"],
        "summary": "Toggle ticket type status",
        "description": "Flips ticket type state between `Active` and `Disabled`.",
        "operationId": "toggleTicketTypeStatus",
        "parameters": [
          {
            "$ref": "#/components/parameters/ticketTypeId"
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket type status toggled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketTypeData"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "API_Key": {
        "type": "apiKey",
        "description": "BTCPay Server Greenfield API Key. Pass as `Authorization: token <key>`. Create under Account > Manage Account > API Keys. Required permission: `btcpay.store.canmodifystoresettings`.",
        "name": "Authorization",
        "in": "header"
      }
    },
    "parameters": {
      "eventId": {
        "name": "eventId",
        "in": "path",
        "description": "The event ID.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "ticketTypeId": {
        "name": "ticketTypeId",
        "in": "path",
        "description": "The ticket type ID.",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "EventNotFound": {
        "description": "Event not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GreenfieldError"
            },
            "example": {
              "code": "event-not-found",
              "message": "The event was not found"
            }
          }
        }
      }
    },
    "schemas": {
...
      "TicketTypeRequest": {
        "type": "object",
        "required": ["name", "price"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Ticket type name (e.g. `VIP`, `Standard`)."
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "description": "Price per ticket. Must be > 0.",
            "example": 50.00
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Description of what this tier includes."
          },
          "quantity": {
            "type": "integer",
            "description": "Number of tickets available. Must be > 0 if the event has maximum capacity. Use `0` for unlimited.",
            "default": 0
          },
          "isDefault": {
            "type": "boolean",
            "default": false,
            "description": "Whether this is the pre-selected default tier. First created type is always default."
          },
          "bundledRaffleId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Open BTCPay Raffle id whose tickets are included with each ticket of this type."
          },
          "bundledRaffleTicketsPerAdmission": {
            "type": "integer",
            "nullable": true,
            "description": "Number of raffle tickets granted per one ticket of this type (0 = none, clears bundle on update). Max 20."
          }
        }
      },
      "TicketTypeData": {
        "type": "object",
        "description": "Full ticket type representation.",
        "properties": {
          "id": {
            "type": "string",
            "description": "Ticket type ID."
          },
          "eventId": {
            "type": "string",
            "description": "Parent event ID."
          },
          "name": {
            "type": "string",
            "description": "Ticket type name."
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "description": "Price per ticket."
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Tier description."
          },
          "quantity": {
            "type": "integer",
            "description": "Total tickets available (0 = unlimited)."
          },
          "quantitySold": {
            "type": "integer",
            "description": "Tickets sold so far."
          },
          "quantityAvailable": {
            "type": "integer",
            "description": "Remaining tickets (quantity - quantitySold)."
          },
          "isDefault": {
            "type": "boolean",
            "description": "Whether this is the default tier."
          },
          "ticketTypeState": {
            "type": "string",
            "enum": ["Active", "Disabled"],
            "description": "Current ticket type state."
          },
          "bundledRaffleId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Open BTCPay Raffle id whose tickets are included with each ticket of this type."
          },
          "bundledRaffleTicketsPerAdmission": {
            "type": "integer",
            "description": "Number of raffle tickets granted per one ticket of this type (0 = none)."
          }
        }
      },
...
      "GreenfieldError": {
        "type": "object",
        "description": "Standard Greenfield API error.",
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "array",
        "description": "Validation error list.",
        "items": {
          "type": "object",
          "properties": {
            "path": {
              "type": "string",
              "description": "Field that caused the error."
            },
            "message": {
              "type": "string",
              "description": "Validation message."
            }
          }
        }
      }
    }
  }
}